+++ include/c_client.h
@@ -69,7 +69,7 @@ SRError DeleteCClient(void** c_client);
 *   \details The final dataset key under which the dataset is stored
 *            is generated from the name that was supplied when the
 *            dataset was created and may be prefixed. See
-*            use_tensor_ensemble_prefix() for more details.
+*            use_dataset_ensemble_prefix() for more details.
 *   \param c_client The client object to use for communication
 *   \param dataset The DataSet object to send
 *   \return Returns SRNoError on success or an error code on failure
@@ -81,7 +81,7 @@ SRError put_dataset(void* c_client, void* dataset);
 *   \details The final dataset key used to locate the dataset
 *            may be formed by applying a prefix to the supplied
 *            name. See set_data_source() and
-*            use_tensor_ensemble_prefix() for more details.
+*            use_dataset_ensemble_prefix() for more details.
 *   \param c_client The client object to use for communication
 *   \param name The name of the dataset object to fetch
 *   \param name_length The length of the name string,
@@ -99,7 +99,7 @@ SRError get_dataset(void* c_client,
 *   \details The old and new dataset keys used to
 *            find and relocate the dataset may be formed by applying
 *            prefixes to the supplied old_name and new_name.
-*            See set_data_source() and use_tensor_ensemble_prefix()
+*            See set_data_source() and use_dataset_ensemble_prefix()
 *            for more details.
 *   \param c_client The client object to use for communication
 *   \param old_name The current name key of the dataset object
@@ -121,7 +121,7 @@ SRError rename_dataset(void* c_client,
 *   \details The source and destination dataset keys used to
 *            locate and store the dataset may be formed by applying
 *            prefixes to the supplied src_name and dest_name.
-*            See set_data_source() and use_tensor_ensemble_prefix()
+*            See set_data_source() and use_dataset_ensemble_prefix()
 *            for more details.
 *   \param c_client The client object to use for communication
 *   \param src_name The source name of the dataset object
@@ -143,7 +143,7 @@ SRError copy_dataset(void* c_client,
 *   \details The dataset key used to locate the dataset to be deleted
 *            may be formed by applying a prefix to the supplied
 *            name. See set_data_source()
-*            and use_tensor_ensemble_prefix() for more details.
+*            and use_dataset_ensemble_prefix() for more details.
 *   \param c_client The client object to use for communication
 *   \param name The name of the dataset object
 *   \param name_length The length of the name string,
@@ -518,7 +518,7 @@ SRError set_model(void* c_client,
                   const char** outputs,
                   const size_t* output_lengths,
                   const size_t n_outputs);
- 
+
  /*!
 *   \brief Set a model (from buffer) in the database for future execution
 *          in a multi-GPU system
@@ -1104,7 +1104,7 @@ SRError model_exists(void* c_client,
 *   \details The dataset key used to check for dataset existence
 *            may be formed by applying a prefix to the supplied
 *            name. See set_data_source()
-*            and use_tensor_ensemble_prefix() for more details.
+*            and use_dataset_ensemble_prefix() for more details.
 *   \param c_client The client object to use for communication
 *   \param name The name of the dataset that will be checked in the database.
 *               The full key corresponding to \p name will be formed
@@ -1200,7 +1200,7 @@ SRError poll_tensor(void* c_client,
 *   \details The dataset key used to check for dataset existence
 *            may be formed by applying a prefix to the supplied
 *            name. See set_data_source()
-*            and use_tensor_ensemble_prefix() for more details.
+*            and use_dataset_ensemble_prefix() for more details.
 *   \param c_client The client object to use for communication
 *   \param name The name of the entity to be checked in the database.
 *               The full key associated to \p name will be formed according
@@ -1249,16 +1249,16 @@ SRError set_data_source(void* c_client,
                         const size_t source_id_length);
 
 /*!
-*   \brief Control whether tensor and dataset names are
-*          prefixed (e.g. in an ensemble) when forming database keys
+*   \brief Control whether tensor names are prefixed (e.g. in an
+*          ensemble) when forming database keys
 *   \details This function can be used to avoid key collisions in an
 *            ensemble by prepending the string value from the
-*            environment variable SSKEYIN to tensor and dataset names.
+*            environment variable SSKEYIN to tensor names.
 *            Prefixes will only be used if they were previously set through
 *            Keys for entities created before this function is called
 *            the environment variables SSKEYOUT and SSKEYIN.
 *            will not be retroactively prefixed.
-*            By default, the client prefixes tensor and dataset keys
+*            By default, the client prefixes tensor keys
 *            with the first prefix specified with the SSKEYIN
 *            and SSKEYOUT environment variables.
 *
@@ -1268,6 +1268,27 @@ SRError set_data_source(void* c_client,
 *   \return Returns SRNoError on success or an error code on failure
 */
 SRError use_tensor_ensemble_prefix(void* c_client, bool use_prefix);
+
+/*!
+*   \brief Control whether dataset names are prefixed (e.g. in an
+*          ensemble) when forming database keys
+*   \details This function can be used to avoid key collisions in an
+*            ensemble by prepending the string value from the
+*            environment variable SSKEYIN to tensor and dataset names.
+*            Prefixes will only be used if they were previously set through
+*            Keys for entities created before this function is called
+*            the environment variables SSKEYOUT and SSKEYIN.
+*            will not be retroactively prefixed.
+*            By default, the client prefixes dataset keys
+*            with the first prefix specified with the SSKEYIN
+*            and SSKEYOUT environment variables.
+*
+*   \param c_client The client object to use for communication
+*   \param use_prefix If true, all future operations on tensors and
+*                     datasets will use a prefix, if available.
+*   \return Returns SRNoError on success or an error code on failure
+*/
+SRError use_dataset_ensemble_prefix(void* c_client, bool use_prefix);
 
 /*!
 *   \brief Control whether model and script names are
@@ -1301,9 +1322,9 @@ SRError use_model_ensemble_prefix(void* c_client, bool use_prefix);
 *            prefixed. By default, the client prefixes aggregation
 *            list keys with the first prefix specified with the SSKEYIN
 *            and SSKEYOUT environment variables.  Note that
-*            use_tensor_ensemble_prefix() controls prefixing
+*            use_dataset_ensemble_prefix() controls prefixing
 *            for the entities in the aggregation list, and
-*            use_tensor_ensemble_prefix() should be given the
+*            use_dataset_ensemble_prefix() should be given the
 *            same value that was used during the initial
 *            setting of the DataSet into the database.
 *   \param c_client The client object to use for communication
@@ -1530,6 +1551,7 @@ SRError get_dataset_list_range(void* c_client, const char* list_name,
                                const size_t list_name_length,
                                const int start_index, const int end_index,
                                void*** datasets, size_t* num_datasets);
+
 /*
 *   \brief Get a range of datasets (by index) from an aggregation list and
            copy them into an already allocated vector of datasets. Note,
@@ -1566,6 +1588,14 @@ SRError _get_dataset_list_range_allocated(void* c_client, const char* list_name,
                                          const size_t list_name_length,
                                          const int start_index, const int end_index,
                                          void** datasets);
+
+
+/*!
+*   \brief Retrieve a string representation of the client
+*   \param c_client The client object to use for communication
+*   \return A string with either the client representation or an error message
+*/
+const char* client_to_string(void* c_client);
 
 #ifdef __cplusplus
 }