+++ src/python/src/pyclient.cpp
@@ -935,6 +935,11 @@ void PyClient::use_tensor_ensemble_prefix(bool use_prefix)
   _client->use_tensor_ensemble_prefix(use_prefix);
 }
 
+void PyClient::use_dataset_ensemble_prefix(bool use_prefix)
+{
+  _client->use_dataset_ensemble_prefix(use_prefix);
+}
+
 void PyClient::use_model_ensemble_prefix(bool use_prefix)
 {
   _client->use_model_ensemble_prefix(use_prefix);
@@ -1350,6 +1355,27 @@ py::list PyClient::get_dataset_list_range(
         // should never happen
         throw SRInternalException("A non-standard exception was encountered "\
                                   "while executing get_dataset_list_range.");
+    }
+}
+
+// Create a string representation of the Client
+std::string PyClient::to_string()
+{
+    try {
+        return _client->to_string();
+    }
+    catch (Exception& e) {
+        // exception is already prepared for caller
+        throw;
+    }
+    catch (std::exception& e) {
+        // should never happen
+        throw SRInternalException(e.what());
+    }
+    catch (...) {
+        // should never happen
+        throw SRInternalException("A non-standard exception was encountered "\
+                                  "while executing to_string.");
     }
 }