+++ doc/changelog.rst
@@ -12,9 +12,14 @@ This section details changes made in the development branch that have not yet be
 
 Description
 
+- Improve documentation of exception handling in Redis server classes
+- Improve error handling for setting of scripts and models
+- Add support to inspect the dimensions of a tensor via get_tensor_dims()
+- Split dataset prefixing control from use_tensor_ensemble_prefix() to use_dataset_ensemble_prefix()
 - Update to the latest version of redis-plus-plus
 - Update to the latest version of PyBind
 - Change documentation theme to sphinx_book_theme and fix doc strings
+- Add print capability for Client and DataSet
 - Add support for inspection of tensors and metadata inside datasets
 - Add support for user-directed logging for Python clients, using Client, Dataset, or LogContext logging methods
 - Add support for user-directed logging for C and Fortran clients without a Client or Dataset context
@@ -29,9 +34,14 @@ Description
 
 Detailed Notes
 
+- Added throw documentation for exception handling in redis.h, redisserver.h, rediscluster.h (PR301_)
+- Added error handling for a rare edge condition when setting scripts and models (PR300_)
+- Added support to inspect the dimensions of a tensor via new get_tensor_dims() method (PR299_)
+- The use_tensor_ensemble_prefix() API method no longer controls whether datasets are prefixed. A new API method, use_dataset_ensemble_prefix() now manages this. (PR298_)
 - Updated from redis-plus-plus v1.3.2 to v1.3.5 (PR296_)
 - Updated from PyBind v2.6.2 to v2.10.3 (PR295_)
 - Change documentation theme to sphinx_book_theme to match SmartSim documentation theme and fix Python API doc string errors (PR294_)
+- Added print capability for Client and DataSet to give details diagnostic information for debugging (PR293_)
 - Added support for retrieval of names and types of tensors and metadata inside datasets (PR291_)
 - Added support for user-directed logging for Python clients via {Client, Dataset, LogContext}.{log_data, log_warning, log_error} methods (PR289_)
 - Added support for user-directed logging without a Client or Dataset context to C and Fortran clients via _string() methods (PR288_)
@@ -48,9 +58,14 @@ Detailed Notes
 - Implemented support for Unix Domain Sockets, including refactorization of server address code, test cases, and check-in tests. (PR252_)
 - A new make target `make lib-with-fortran` now compiles the Fortran client and dataset into its own library which applications can link against (PR245_)
 
+.. _PR301: https://github.com/CrayLabs/SmartRedis/pull/301
+.. _PR300: https://github.com/CrayLabs/SmartRedis/pull/300
+.. _PR299: https://github.com/CrayLabs/SmartRedis/pull/299
+.. _PR298: https://github.com/CrayLabs/SmartRedis/pull/298
 .. _PR296: https://github.com/CrayLabs/SmartRedis/pull/296
 .. _PR295: https://github.com/CrayLabs/SmartRedis/pull/295
 .. _PR294: https://github.com/CrayLabs/SmartRedis/pull/294
+.. _PR293: https://github.com/CrayLabs/SmartRedis/pull/293
 .. _PR291: https://github.com/CrayLabs/SmartRedis/pull/291
 .. _PR289: https://github.com/CrayLabs/SmartRedis/pull/289
 .. _PR288: https://github.com/CrayLabs/SmartRedis/pull/288